requirements/requirements_docs.txt
ansys-sphinx-theme==0.8.0
imageio==2.24.0
imageio-ffmpeg==0.4.8
ipython==8.8.0
ipywidgets==8.0.4
joblib==1.2.0
jupyterlab==3.5.2
matplotlib==3.6.3
nbsphinx==0.8.12
numpydoc==1.5.0
osmnx
pypandoc==1.10
pytest-sphinx==0.5.0
pyvista==0.37.0
recommonmark==0.7.1
scikit-learn==1.2.0
Sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinx-autodoc-typehints==1.21.4
sphinx-copybutton==0.5.1
sphinx-gallery==0.11.1
sphinx-notfound-page==0.8.3
sphinxcontrib-websupport==1.2.4
SRTM.py
utm
